End-to-End Integration: Embeddings Upload to Quantization to Indexing (#148)
* feat(versioning): use `SipHasher24` for hashing versions

* feat(buffered_io): use a function to map `Hash` to file path

* chore: add `RawVectorEmbedding` and rename `VectorEmbedding` to `QuantizedVectorEmbedding`

* fix(versioning): use little endian for (de)serialization

* chore(types): chage `current_version` on `VectorStore` from `ArcShift<Option<Hash>>` to `ArcShift<Hash>`

* chore(api_service): flush `BufferManager` after serialization

* feat: save raw embeddings in different files by version

* fix(vector_store): fix indexing

* feat(vector_store/indexing): initailize new nodes with appropriate version

* fix(api_service): flush all `BufferManager`s after committing transaction

* chore(api_serive): return a `Result` from `run_upload`

* chore: remove unused `CustomBufferedReader` and `CustomBufferedWriter`

* feat: trait quantization metric with the first batch of vectors

* fix(quantization): fix `quantize_to_u8_bits` function

* feat: infer storage type from dimension

* chore: improve todo comments

* chore: remove unused `vector` module from `serializer`

* using vectorId strict type instead of str

* adding a new vectors error variant

* added a function to return a raw embedding vector by its id from a vector store, and its documentation

* added new error variants to VectorsError

* created a vector repo method to return a raw vector by its id and a vector store id

* feat: auto config storage type

* feat: implement some `unimplemented!` distance functions

* fix(dot_product): improve octal dot product

* feat(api/vectordb/indexes): partially integrate index creation API

* feat(quantization): integrate product quantization

* chore: revert product quantization changes

* feat: use number of clusters to configure storage type

* feat(api_service/run_upload): track embedding offset for re-indexing

* fix(indexing): use same file offset for props of all HNSW levels of a single node

* feat(api): integrate index creation API

* fix: typos

* feat(api/vectordb/vectors): integrate vector delete API

* feat(api/vectordb/transactions): integrate vectors delete API

* feat: use SIMD dot product for binary and quaternary

* chore: add test for SIMD binary dot product correctness

* refactor: `*_simple` -> `*_scalar` for non SIMD functions

* feat(dot_product): integrate SIMD dot product for octal

---------

+++ src/storage/mod.rs
@@ -27,7 +27,7 @@ pub enum Storage {
         quant_vec: Vec<u8>,
     },
     SubByte {
-        mag: u32,
+        mag: f32,
         quant_vec: Vec<Vec<u8>>,
         resolution: u8,
     },
